requirements.txt
asgiref==3.7.2
autopep8==2.0.4
backports.zoneinfo==0.2.1
black==24.2.0
click==8.1.7
Django==4.2.10
django-cors-headers==4.3.1
django-filter==23.5
django-silk==5.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==23.3.0
gprof2dot==2022.7.29
inflection==0.5.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pycodestyle==2.11.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
uritemplate==4.1.1
